Add API to invoke parol from build scripts
Right now parol has a great command line interface,
but is extremely difficult to invoke from build.rs

This unifies the logic for the CLI with a new API
for invoking parol from build scripts.

The design is based off the APIs for the bindgen and cc crates,
which are the main build-script APIs I know of.